+++ index.js
@@ -1,30 +1,32 @@
 export async function waitForAnimation(element, options = {}) {
-  if (!element.getAnimations) {
+  if (typeof element.getAnimations !== 'function') {
     return;
   }
 
-  await waitForFrame();
+  if (getAnimations().length < 1 && options.maybe !== true) {
+    await waitForStart();
+  }
 
   return Promise.allSettled(
-    element
-      .getAnimations({
-        subtree: options.subtree
-      })
-      .filter((animation) => {
-        if (options.transitionProperty) {
-          return animation.transitionProperty === options.transitionProperty;
-        }
-
-        if (options.animationName) {
-          return animation.animationName === options.animationName;
-        }
-
-        return true;
-      })
-      .map((animation) => animation.finished)
+    getAnimations().map((animation) => animation.finished)
   );
-}
 
-export function waitForFrame() {
-  return new Promise(window.requestAnimationFrame);
+  function waitForStart() {
+    return new Promise((resolve) => {
+      element.addEventListener('animationstart', resolve, { once: true });
+      element.addEventListener('transitionstart', resolve, { once: true });
+    });
+  }
+
+  function getAnimations() {
+    return element
+      .getAnimations({ subtree: options.subtree })
+      .filter((animation) => {
+        return options.transitionProperty
+          ? animation.transitionProperty === options.transitionProperty
+          : options.animationName
+          ? animation.animationName === options.animationName
+          : true;
+      });
+  }
 }